Split off activity_ptr into its own header
